refactor: move threat hunting constants to a dedicated file for better organization
