+++ plugins/wazuh-analysis/public/groups/threat-intelligence/applications.ts
@@ -5,8 +5,6 @@ import {
   AppUpdater,
 } from '../../../../../src/core/public';
 import {
-  THREAT_HUNTING_ID,
-  THREAT_HUNTING_TITLE,
   VULNERABILITY_DETECTION_ID,
   VULNERABILITY_DETECTION_TITLE,
 } from './constants';
@@ -14,6 +12,10 @@ import {
   MITRE_ATTACK_ID,
   MITRE_ATTACK_TITLE,
 } from './apps/mitre-att&ck/constants';
+import {
+  THREAT_HUNTING_ID,
+  THREAT_HUNTING_TITLE,
+} from './apps/threat-hunting/constants';
 
 export function getThreatIntelligenceApps(updater$?: Subject<AppUpdater>) {
   return [
