+++ plugins/wazuh-analysis/public/groups/threat-intelligence/apps/threat-hunting/application.tsx
@@ -4,8 +4,9 @@ import ReactDOM from 'react-dom';
 import { ThreatIntelligenceNavGroup } from '../..';
 import { Layout } from '../../../layout';
 import { createSideNavItems } from '../../../side-nav';
-import { THREAT_HUNTING_ID, THREAT_INTELLIGENCE_TITLE } from '../../constants';
+import { THREAT_INTELLIGENCE_TITLE } from '../../constants';
 import { ThreatHuntingApp } from './threat-hunting-app';
+import { THREAT_HUNTING_ID } from './constants';
 
 export const renderApp = async (params: AppMountParameters) => {
   const items = createSideNavItems({
