+++ plugins/wazuh-analysis/public/groups/threat-intelligence/apps/threat-hunting/threat-hunting-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { THREAT_HUNTING_TITLE } from '../../constants';
+import { THREAT_HUNTING_TITLE } from './constants';
 
 interface ThreatHuntingAppProps {
   params: AppMountParameters;
